requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
sqlalchemy==2.0.28
pymysql==1.1.0
pandas==2.2.1
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
faiss-cpu==1.7.4 